fix: combine map and resultsmap into one component
